fix: remove image strings from user prompts, remove unused logger
before sending the prompt to the backend, remove any images encoded in a base64 string, which are indicated by markdown image links like ![alt text here](data:image/png.*). This significantly reduces the number of tokens sent to the backend. Previously the long string containing the image would often result in the request to the backend failing, either because it would exceed the maximum request length of the httpx client or the token limit of the backend